mythril-tests/requirements.txt
asn1crypto==1.5.1
bitarray==2.9.2
blake2b-py==0.3.0
cached-property==1.5.2
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
ckzg==1.0.2
coincurve==20.0.0
coloredlogs==15.0.1
configparser==7.0.0
contourpy==1.2.1
coverage==6.5.0
cycler==0.12.1
cytoolz==0.12.3
dictionaries==0.0.2
distlib==0.3.8
eth-abi==4.2.1
eth-account==0.11.2
eth-bloom==1.0.4
eth-hash==0.3.3
eth-keyfile==0.8.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
ethereum_input_decoder==0.2.2
filelock==3.15.4
fonttools==4.53.1
hexbytes==0.2.3
humanfriendly==10.0
identify==2.6.0
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lru-dict==1.3.0
MarkupSafe==2.0.1
matplotlib==3.9.1
mock==5.1.0
mypy_extensions==0.4.4
mythril==0.24.8
nodeenv==1.9.1
numpy==2.0.0
packaging==24.1
parsimonious==0.9.0
persistent==6.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==2.20.0
py-ecc==1.4.7
py-evm==0.7.0a1
py-flags==1.1.4
py-solc==3.2.0
py-solc-x==1.1.1
pycparser==2.22
pycryptodome==3.20.0
pyethash==0.1.27
pyparsing==2.4.7
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rlp==3.0.0
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
setuptools==70.3.0
six==1.16.0
sortedcontainers==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
transaction==4.0
trie==2.2.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
z3-solver==4.12.5.0
zope.deferredimport==5.0
zope.interface==6.4.post2
zope.proxy==5.2
